requirements.txt
great_expectations
python-lambda-local
ruamel.yaml
rx==1.6.1
pandas
